Pins have been moved to the pins folder and tests have been moved as well accordingly for pins and brainzplayer.
