+++ listenbrainz/webserver/static/js/src/RecentListens.tsx
@@ -28,8 +28,8 @@ import BrainzPlayer from "./brainzplayer/BrainzPlayer";
 import ErrorBoundary from "./ErrorBoundary";
 import ListenCard from "./listens/ListenCard";
 import Loader from "./components/Loader";
-import PinRecordingModal from "./PinRecordingModal";
-import PinnedRecordingCard from "./PinnedRecordingCard";
+import PinRecordingModal from "./pins/PinRecordingModal";
+import PinnedRecordingCard from "./pins/PinnedRecordingCard";
 import {
   formatWSMessageToListen,
   getPageProps,
