+++ listenbrainz/webserver/static/js/tests/pins/UserPins.test.tsx
@@ -3,16 +3,16 @@
 import * as React from "react";
 import { mount } from "enzyme";
 import * as timeago from "time-ago";
-import { GlobalAppContextT } from "./GlobalAppContext";
-import APIService from "./APIService";
-import BrainzPlayer from "./brainzplayer/BrainzPlayer";
+import { GlobalAppContextT } from "../../src/GlobalAppContext";
+import APIService from "../../src/APIService";
+import BrainzPlayer from "../../src/brainzplayer/BrainzPlayer";
 
-import * as pinsPageProps from "../tests/__mocks__/userPinsProps.json";
-import * as APIPins from "../tests/__mocks__/pinProps.json";
+import * as pinsPageProps from "../__mocks__/userPinsProps.json";
+import * as APIPins from "../__mocks__/pinProps.json";
 
-import { getListenablePin } from "./utils";
+import { getListenablePin } from "../../src/utils";
 
-import UserPins from "./UserPins";
+import UserPins from "../../src/pins/UserPins";
 
 // Font Awesome generates a random hash ID for each icon everytime.
 // Mocking Math.random() fixes this
